requirements.txt
alabaster==1.0.0
altgraph==0.17.4
appdirs==1.4.4
argcomplete==3.5.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.16.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2025.4
cryptography==45.0.3
daphne==4.2.0
Django==5.1.1
django-crispy-forms==2.4
django-extensions==3.2.3
docutils==0.21.2
Faker==30.3.0
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
importlib_resources==6.4.5
incremental==24.7.2
Jinja2==3.1.5
libusb1==3.1.0
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pefile==2023.2.7
pillow==10.4.0
playsound==1.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyOpenSSL==25.1.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-escpos==3.1
pyusb==1.2.1
pywebview>=3.6
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.0
redis==6.1.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
zope.interface==7.2
python-dotenv==1.1.1
psycopg2-binary==2.9.10
whitenoise==6.7.0
beautifulsoup4==4.12.3
django-colorfield==0.14.0